Django-Image-Uploader-Web-Application/requirements.txt
absl-py==2.1.0
addict==2.4.0
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
autopep8==2.0.2
backports.zoneinfo==0.2.1
basicsr==1.4.2
cachetools==5.3.2
certifi==2024.2.2
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.0.3
colorama==0.4.6
cryptography==42.0.3
cssselect2==0.7.0
cycler==0.10.0
defusedxml==0.7.1
Django==3.2.22
django-allauth==0.61.1
django-countries==7.5.1
django-crispy-forms==2.0
django-utils==0.0.2
django-widget-tweaks==1.4.8
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
Flask==0.12.2
future==0.18.3
google-auth==2.28.0
google-auth-oauthlib==0.4.6
grpcio==1.60.1
html5lib==1.1
idna==3.6
imageio==2.31.2
importlib-metadata==6.7.0
itsdangerous==2.0.1
Jinja2==3.0.3
kafka-python==2.0.2
kiwisolver==1.3.1
lmdb==1.4.1
lxml==5.1.0
Markdown==3.4.4
MarkupSafe==2.1.5
matplotlib==3.4.2
mysql-connector-python==8.0.27
networkx==2.6.3
numpy==1.20.3
oauthlib==3.2.2
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
oscrypto==1.3.0
packaging==23.2
pep8==1.7.1
phonenumbers==8.12.24
Pillow==9.5.0
platformdirs==4.0.0
playsound==1.3.0
protobuf==3.20.3
psycopg2==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser==2.21
pyHanko==0.17.2
pyhanko-certvalidator==0.20.1
PyJWT==2.8.0
PyMySQL==1.0.2
pyparsing==2.4.7
pypdf==3.7.0
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.1
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.3.0
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.1.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scipy==1.7.3
six==1.16.0
sqlparse==0.3.1
stripe==5.4.0
svglib==1.5.1
tb-nightly==2.12.0a20230113
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tifffile==2021.11.2
tinycss2==1.2.1
tomli==2.0.1
torch==1.13.1
torchvision==0.14.1
tqdm==4.66.2
typing_extensions==4.7.1
tzdata==2024.1
tzlocal==4.3
uritools==4.0.1
urllib3==1.26.15
webencodings==0.5.1
Werkzeug==2.0.2
xhtml2pdf==0.2.9
yapf==0.40.2
zipp==3.6.0